refactor(threading): move threading classes to top-level
